Moved fetch/toggle logic for hiding to repo Fixed sorting
