+++ app/src/main/java/com/topjohnwu/magisk/di/RepositoryModule.kt
@@ -7,7 +7,7 @@ import org.koin.dsl.module
 
 
 val repositoryModule = module {
-    single { MagiskRepository(get(), get()) }
+    single { MagiskRepository(get(), get(), get()) }
     single { ModuleRepository(get(), get(), get(), get()) }
     single { LogRepository(get()) }
 }
